feat: v0.1.49 - fix combined nullable bug
