+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "tspec",
   "description": "Generate OpenAPI 3.0 spec from TypeScript code.",
-  "version": "0.1.47",
+  "version": "0.1.49",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "license": "MIT",
