+++ src/generator/types.ts
@@ -1,0 +1,4 @@
+import { OpenAPIV3 } from 'openapi-types';
+
+export type Schema = OpenAPIV3.SchemaObject | OpenAPIV3.ReferenceObject;
+export type SchemaMapping = Record<string, OpenAPIV3.SchemaObject>;
